Fix Obstacle Class const to float

+++ src/gameObjects/Obstacle.h
@@ -8,8 +8,8 @@ private:
 	bool jumped;
 	Texture2D texture;
 	float scale;
-	const int SPAWN_POSITION_X = 1280;
-	const int SPAWN_POSITION_y = 600;
+	const float SPAWN_POSITION_X = 1280.0f;
+	const float SPAWN_POSITION_y = 600.0f;
 	const int MAX_GRAVITY_FORCE = 200;
 
 public:
